% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gwdegree}
\alias{gwdegree}
\title{Run GW-Degree Shiny Application}
\usage{
gwdegree()
}
\value{
NULL Called for side-effect of launching shiny application
}
\description{
Run GW-Degree Shiny Application
}
\examples{
\dontrun{
gwdegree()
}
}

